test: remove proxy host from both raw & normalized fixtures
